% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN-AFTplot.R
\name{AFTplot}
\alias{AFTplot}
\title{Q-Q-plot of Predicted Survival Times}
\usage{
AFTplot(model)
}
\arguments{
\item{model}{An object of class \code{\link[survival]{survreg}}.}
}
\description{
Accelerated Failure Time models are based on the assumption that different levels
 of a predictor (e. g. different treatments) have a multiplicative effect on the
 survival time. Plotting the predicted survival time quantiles of different
 predictor levels against each other allows to assess the validity of the assumption.
 If the assumption is valid, the quantile pairs will approximately lie on a straight
 line through the origin.
}
\note{
If the \code{model} includes frailties, they have to be estimated with
 \code{sparse = FALSE} to allow predictions.
}
\references{
M. J. Bradburn, T. G. Clark, S. B. Love, und D. G. Altman, „Survival analysis
 Part III: multivariate data analysis -- choosing a model and assessing its
 adequacy and fit“, Br. J. Cancer, Bd. 89, Nr. 4, S. 605–611, Aug. 2003.

W. R. Swindell, „ACCELERATED FAILURE TIME MODELS PROVIDE A USEFUL STATISTICAL
 FRAMEWORK FOR AGING RESEARCH“, Exp Gerontol, Bd. 44, Nr. 3, S. 190–200, März 2009.
}

