% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN-compare_PSMdist.R
\name{compare_PSMdist}
\alias{compare_PSMdist}
\title{Compare Different Distributions for Parametric Survival Models}
\usage{
compare_PSMdist(Formula, Data, Distributions = c("weibull", "exponential",
  "lognormal", "loglogistic"), parallel = TRUE, cpus = NULL,
  type = "SOCK", level = 0.95, ...)
}
\arguments{
\item{Formula}{A formula expression as for other regression models passed to
\code{\link[survival]{survreg}}. The response has to be a survival object
as returned by the \code{\link[survival]{Surv}} function.}

\item{Data}{A data frame where the variables in formula are found.}

\item{Distributions}{A character vector. The names of the distributions to be
used. See \code{\link[survival]{survreg}} for the possible distributions.}

\item{parallel}{Logical determinating parallel or sequential execution, passed
to \code{\link[snowfall]{sfInit}}.}

\item{cpus}{Passed to \code{\link[snowfall]{sfInit}}.}

\item{type}{Passed to \code{\link[snowfall]{sfInit}}.}

\item{level}{Passed to \code{\link[snowfall]{sfInit}}.}

\item{...}{Further arguments passed to \code{\link[snowfall]{sfInit}},
\code{\link{confint}}, and \code{\link[survival]{survreg}}.}
}
\value{
A list consiting of:
\describe{
 \item{\code{Formula}}{The formula used for the fitting call.}
 \item{\code{Distribution}}{A character vector naming the distributions used
   for the models. It defines the order for the following list elements.}
 \item{\code{Fit}}{A logical vector: if \code{FALSE} the corresponding model
   could not be fitted without problems and a \code{Message} is returned.}
 \item{\code{Message}}{A character vector: \code{NA} if the corresponding model
   was fitted without problems, otherwise the error message.}
 \item{\code{AIC}}{A numeric vector with the \code{\link[stats]{AIC}} of the
   models. \code{NA} if the fitting of the corresponding model caused problems.}
 \item{\code{Effects}}{A numeric array: the levels of the model effects define
   the rows, three columns (estimated coefficient, lower and upper confidence
   limits at 95\%-level as default), and separate layers for each distribution.
   \code{NA} if the fitting of the corresponding model caused problems.}
 \item{\code{logScale}}{A numeric matrix with one row for each distribution,
   and three columns (estimated coefficient, lower and upper confidence limits
   at 95\%-level as default). \code{NA} if the fitting of the corresponding
   model caused problems.}
 \item{\code{FrailtyVar}}{A numeric vector with the estimated variance of the
   frailty effect (if one is included.). \code{NA} if the fitting of the
   corresponding model caused problems.}
}
}
\description{
Choosing the best distribution for parametric survivial models is not trivial.
One possibility is to select of all plausible distributions the one with which
the resulting models' AIC is minimal. Therefore, several models have to be
fitted and compared. \code{compare_PSMdist} fits the models for all distributions
that come into consideration by calling \code{\link[survival]{survreg}}. The
fitting of the models is parallelised via \pkg{\link{snowfall}}.
}
\details{
If \code{parallel = TRUE} (the default) and \code{cpus}
 is not specified, on Unix systems all but one CPU will be used by
 \code{\link[snowfall]{sfInit}}. Other systems will use \code{cpus = 2}.
}
\examples{
\donttest{
intS2 <- with(MIC, create_int2Surv(concentration, inhibition))
compare_PSMdist(as.formula("intS2 ~ region"), Data = cbind(intS2, MIC))
compare_PSMdist(as.formula("intS2 ~ region + frailty(herd, sparse = FALSE)"),
 Data = cbind(intS2, MIC), cpus = 2, control = survreg.control(maxiter = 100))
compare_PSMdist(as.formula("intS2 ~ 1 + frailty(herd, sparse = FALSE)"),
 Data = cbind(intS2, MIC), cpus = 2, control = survreg.control(maxiter = 100))
}

}
\seealso{
\code{\link{show_comparison}} produces a print- and readable table of
 the comparison.
}

