% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN-create_int2Surv.R
\name{create_int2Surv}
\alias{create_int2Surv}
\title{Create an interval censored Survival object}
\usage{
create_int2Surv(obsTimes, ev, defTimes = unique(obsTimes))
}
\arguments{
\item{obsTimes}{A numeric vector with the observed survival times (the right
interval bounds.)}

\item{ev}{Were events observed at \code{obsTimes}? A vector of the same length
as \code{obsTimes}, either \code{logical} oder \code{integer} with 0 = censored
and 1 = event.}

\item{defTimes}{A numeric vector with all observation times.}
}
\value{
An object of class \code{\link[survival]{Surv}} with \code{type =
 "interval2"}.
}
\description{
Create a survival object from interval censored data with fixed observation
 times where only times with events or the last observation times (leading
 to right censoring) were reported. Observation times were defined before
 conducting the study and that were the same for all subjects.
}
\examples{
with(MIC, create_int2Surv(concentration, inhibition))

}

