% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUN-CSRplot.R
\name{CSRplot}
\alias{CSRplot}
\title{Cox-Snell-Residual Plot}
\usage{
CSRplot(model)
}
\arguments{
\item{model}{An object of class \code{\link[survival]{survreg}}.}
}
\description{
Plots the Cox-Snell-Residuals of a model fitted using \code{\link[survival]{survreg}}
 against the Kaplan-Meier estimator of their cumulative hazard.
}
\note{
If the \code{model} includes frailties, they have to be estimated with
 \code{sparse = FALSE} to allow the residuals to be extracted.
}
\references{
P. Glomb, „Statistische Modelle und Methoden in der Analyse von Lebenszeitdaten“,
 Diplomarbeit, Carl-von-Ossietzky-Universität, Oldenburg, 2007.

C. P. Farrington, ‘Residuals for Proportional Hazards Models with Interval-Censored
 Survival Data’, Biometrics, vol. 56, no. 2, pp. 473–482, 2000.
}

