% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN-Iplot.R
\name{Iplot}
\alias{Iplot}
\title{Influential Plot}
\usage{
Iplot(model, limit = 10)
}
\arguments{
\item{model}{An object of class \code{\link[survival]{survreg}}.}

\item{limit}{Relative parameter changes above \code{limit}\% are marked in red.}
}
\description{
Plot of the relative parameter change due to addition of each observation. In
 fact, it nicely plots \code{\link[survival]{residuals.survreg}(model, type = "dfbeta")}
 for each fixed effect separately.
}
\note{
If the \code{model} includes frailties, they have to be estimated with
 \code{sparse = FALSE} to allow the residuals to be extracted.
}
\references{
W. R. Swindell, „ACCELERATED FAILURE TIME MODELS PROVIDE A USEFUL STATISTICAL
 FRAMEWORK FOR AGING RESEARCH“, Exp Gerontol, Bd. 44, Nr. 3, S. 190–200, März 2009.
}

