% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DOC-MIC.R
\docType{data}
\name{MIC}
\alias{MIC}
\title{Minimal Inhibitory Concentrations of Cefalonium against Staphylococcus aureus}
\format{A data frame with 609 row and 6 variables. The variables are as follows:

\itemize{
  \item{\code{region}        : The part of Germany where the sample was collected.}
  \item{\code{herd}          :The herd where the sample was collected.}
  \item{\code{cow}           : The cow ... .}
  \item{\code{sample}        : Identification of the quarter milk sample.}
  \item{\code{concentration} : The minimal concentration that inhibited the growth of
   the S. aureus isolate or the highest tested concentration.}
  \item{\code{inhibition}    : Was inhibition of this S. aureus isolate observed?}
}}
\source{
Knorr, N; Zoche-Golob, V; Krömker, V. Determination of Minimal
 Inhibitory Concentrations of Cephalonium, Cephapirin and Cefquinome Against
 Bovine Mastitis Pathogens From Subclinical Mastitis Considering Regional
 Variances Within the Scope of a Germany-wide Study. [Article in preparation]
}
\usage{
MIC
}
\description{
Minimal Inhibitory Concentrations (MIC) of Cefalonium against Staphylococcus (S.) aureus
 from bovine subclinical mastitis. The inhibitions of the S. aureus isolates were
 tested at concentrations of 0.125, 0.25, 0.5, 1.0, 2.0, 4.0, 8.0, 16.0, 32.0, and 64.0
 \eqn{\mu}g/ml of Cefalonium.
}
\keyword{datasets}

