% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN-NPplot.R
\name{NPplot}
\alias{NPplot}
\title{Graphic Comparison of Parametric Survival Models to Correponding Non-Parametric Models}
\usage{
NPplot(model, data)
}
\arguments{
\item{model}{An object of class \code{\link[survival]{survreg}}.}

\item{data}{The data used to fit \code{model}.}
}
\value{
One plot per predictor level (if more than one,
 \code{\link[graphics]{par}(mfrow = c(2, 2))} is used.)
}
\description{
Plots of the predicted survival of \code{model} and of corresponding non-parametric
 models.
}
\details{
\code{\link{predict_survreg}} is used internally to calculate the predicted
 survival for \code{model}. As non-parametric references, the Kaplan-Meier (for
 right censored data) and the Turnbull (for interval censored data) models are
 used by calling \code{\link[survival]{survfit}}.
 Frailties and interactions are ignored for the non-parametric predictions.
}
\note{
Currently only right- and interval-censored data have been tested.
}
\examples{
\donttest{
intS2 <- with(MIC, create_int2Surv(concentration, inhibition))
psm1 <- survival::survreg(as.formula("intS2 ~ region +
 frailty(herd, sparse = FALSE)"), data = cbind(intS2, MIC))
NPplot(psm1, cbind(intS2, MIC))

psm2 <- survival::survreg(as.formula("Surv(concentration, inhibition) ~ region +
 frailty(herd, sparse = FALSE)"), data = MIC)
NPplot(psm2, MIC)

rm(psm1, psm2, intS2)
}

}
\references{
M. Fay, P. Shaw, „Exact and Asymptotic Weighted Logrank Tests for Interval
 Censored Data: The interval R Package“, Journal of Statistical Software, Bd. 36,
 Nr. 2, S. 1-34, 2010.

K. Goethals, B. Ampe, D. Berkvens, H. Laevens, P. Janssen, und L. Duchateau,
 „Modeling interval-censored, clustered cow udder quarter infection times
 through the shared gamma frailty model“, JABES, Bd. 14, Nr. 1, S. 1–14, März 2009.

W. R. Swindell, „ACCELERATED FAILURE TIME MODELS PROVIDE A USEFUL STATISTICAL
 FRAMEWORK FOR AGING RESEARCH“, Exp Gerontol, Bd. 44, Nr. 3, S. 190–200, März 2009.
}

