% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FUN-predict_survreg.R
\name{predict_survreg}
\alias{predict_survreg}
\title{Predictions from Parametric Survival Models}
\usage{
predict_survreg(mod, type = "quantile", quantiles = c(0.5, 0.9),
  times = NULL, conf.int = FALSE)
}
\arguments{
\item{mod}{An object of class \code{\link[survival]{survreg}}.}

\item{type}{A character string: one of \code{c("quantile", "survival")}.}

\item{quantiles}{A numeric vector of survival quantiles for which the times
are predicted.}

\item{times}{A numeric vector of survival times for which the survival percentage
should be predicted.}

\item{conf.int}{Logical. Shall the confidence intervals be predicted as well?
 (currently not supported)}
}
\value{
For \code{type = survival} a data frame with one row per predictor level
 and one column per requested time.

For \code{type = quantile} a data frame with one row per predictor level
 and one column per requested quantile.
}
\description{
Predicts survival or quantiles from Parametric Survival Models that were fitted
 using \code{\link[survival]{survreg}}. Unlike with \code{\link[survival]{predict.survreg.penal}},
 predictions are population averages (i.e. the frailty effect is set to its
 \code{mean = 0}).
}
\note{
Predictions are returned for all levels of all predictors (quantitative
 predictors not tested yet).
}
\examples{
intS2 <- with(MIC, create_int2Surv(concentration, inhibition))
psm <- survival::survreg(as.formula("intS2 ~ region +
 frailty(herd, sparse = FALSE)"), data = cbind(intS2, MIC))
predict_survreg(psm)
predict_survreg(psm, type = "survival", times = c(0.5, 1))
rm(psm, intS2)
}

