% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN-show_comparison.R
\name{show_comparison}
\alias{show_comparison}
\title{Printable Comparison of Different Distributions for Parametric Survival Models}
\usage{
show_comparison(comp)
}
\arguments{
\item{comp}{A list returned by \code{\link{compare_PSMdist}}.}
}
\description{
Printable Comparison of Different Distributions for Parametric Survival Models
}
\examples{
intS2 <- with(MIC, create_int2Surv(concentration, inhibition))
dd <- compare_PSMdist(as.formula("intS2 ~ region"), Data = cbind(intS2, MIC))
show_comparison(dd)
rm(intS2, dd)

}

